msm8937-common: test 2
unmount /vendor and mout again to prevent device is busy, and maybe fix compass in some devices

+++ configs/check_features.sh
@@ -3,12 +3,9 @@
 device=`getprop ro.boot.device`
 radio=`getprop ro.boot.radio`
 
-mount /dev/block/bootdevice/by-name/oem /vendor
+umount /vendor
 
-if [ "$device" != "cedric" ] && [ "$device" != "owens" ] && [ "$device" != "perry" ]; then
-    # cedric, owens & perry have e-compasses
-    rm /vendor/etc/permissions/android.hardware.sensor.compass.xml
-fi
+mount /dev/block/bootdevice/by-name/oem /vendor
 
 if [ "$device" == "montana" ]; then
     if [ "$radio" != "APAC" ] && [ "$radio" != "EMEA" ]; then
@@ -18,8 +15,8 @@ if [ "$device" == "montana" ]; then
         rm /vendor/etc/permissions/android.hardware.nfc.hce.xml
         rm -r /system/app/NfcNci
     fi
-    if [ "$radio" == "APAC" ] || [ "$radio" == "EMEA" ]; then
-        # On montana APAC and EMEA radios don't have e-compasses
+    if [ "$radio" == "APAC" ] || [ "$radio" == "EMEA" ] && [ "$device" != "cedric" ] && [ "$device" != "owens" ] && [ "$device" != "perry" ]; then
+        # On montana APAC and EMEA radios don't have e-compasses and cedric, owens & perry have e-compasses
         rm /vendor/etc/permissions/android.hardware.sensor.compass.xml
     fi
 fi
